+++ modules/elementary_functions/src/cpp/DotPower.cpp
@@ -16,145 +16,123 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // LICENCE_BLOCK_END
 //=============================================================================
+#include <limits>
+#include <complex>
 #include "DotPower.hpp"
 #include "MatrixCheck.hpp"
 #include "complex_abs.hpp"
 //=============================================================================
 namespace Nelson {
-/**
- * These are the power functions...
- */
-// cicPower --> pow_zi
-// cfcPower --> pow_zz
-// zdzPower --> pow_zz
-// cccPower --> pow_zz
-// zzzPower --> pow_zz
-// zizPower --> pow_zi
-// didPower --> pow_di
-// dddPower --> pow_dd
-// fifPower --> pow_ri
-// fffPower --> pow_dd
-
-template <class T>
+//=============================================================================
 void
-complex_divide(T* c, const T* a, const T* b)
+power_zi(double* p, const double* a, int b)
 {
-    double ratio, den;
-    double abr, abi, cr;
-    if ((abr = b[0]) < 0.) {
-        abr = -abr;
-    }
-    if ((abi = b[1]) < 0.) {
-        abi = -abi;
-    }
-    if (abr <= abi) {
-        if (abi == 0) {
-            if (a[1] != 0 || a[0] != 0) {
-                abi = 1.;
-            }
-            c[1] = c[0] = (float)(abi / abr);
-            return;
-        }
-        ratio = b[0] / b[1];
-        den = b[1] * (1 + ratio * ratio);
-        cr = (a[0] * ratio + a[1]) / den;
-        c[1] = (T)((a[1] * ratio - a[0]) / den);
-    } else {
-        ratio = b[1] / b[0];
-        den = b[0] * (1 + ratio * ratio);
-        cr = (a[0] + a[1] * ratio) / den;
-        c[1] = (T)((a[1] - a[0] * ratio) / den);
-    }
-    c[0] = (T)(cr);
+    std::complex<double> ca(a[0], a[1]);
+    std::complex<double> cc = std::pow(ca, b);
+    p[0] = cc.real();
+    p[1] = cc.imag();
 }
 //=============================================================================
 void
-power_zi(double* p, double* a, int b) /* p = a**b  */
+power_zz(double* c, const double* a, const double* b)
 {
-    int n;
-    unsigned long u;
-    double t;
-    double q[2], x[2];
-    static double one[2] = { 1.0, 0.0 };
-    n = b;
-    q[0] = 1;
-    q[1] = 0;
-    if (n == 0) {
-        goto done;
+    double mag = complex_abs<double>(a[0], a[1]);
+    if (mag == 0) {
+        c[0] = 0;
+        c[1] = 0;
+        return;
     }
-    if (n < 0) {
-        n = -n;
-        complex_divide<double>(x, one, a);
+    std::complex<double> ca(a[0], a[1]);
+    std::complex<double> cb(b[0], b[1]);
+    std::complex<double> cc;
+    if (ca.imag() == 0) {
+        cc = pow(ca.real(), cb);
     } else {
-        x[0] = a[0];
-        x[1] = a[1];
+        cc = std::exp(cb * log(ca));
     }
-    for (u = n;;) {
-        if (u & 01) {
-            t = q[0] * x[0] - q[1] * x[1];
-            q[1] = q[0] * x[1] + q[1] * x[0];
-            q[0] = t;
-        }
-        if (u >>= 1) {
-            t = x[0] * x[0] - x[1] * x[1];
-            x[1] = 2 * x[0] * x[1];
-            x[0] = t;
-        } else {
-            break;
-        }
+    c[0] = cc.real();
+    if (ca.imag() == 0 && !std::isfinite(cc.imag())) {
+        c[1] = 0;
+    } else {
+        c[1] = cc.imag();
     }
-done:
-    p[1] = q[1];
-    p[0] = q[0];
-}
-//=============================================================================
-void
-power_zz(double* c, double* a, double* b)
-{
-    double logr, logi, x, y;
-    logr = log(complex_abs<double>(a[0], a[1]));
-    logi = atan2(a[1], a[0]);
-    x = exp(logr * b[0] - logi * b[1]);
-    y = logr * b[1] + logi * b[0];
-    c[0] = x * cos(y);
-    c[1] = x * sin(y);
 }
 //=============================================================================
 double
 power_di(double a, int b)
 {
-    double pow, x;
-    int n;
-    pow = 1;
-    x = a;
-    n = b;
-    if (n != 0) {
-        if (n < 0) {
-            n = -n;
-            x = 1 / x;
-        }
-        for (unsigned long u = n;;) {
-            if (u & 01) {
-                pow *= x;
-            }
-            if (u >>= 1) {
-                x *= x;
-            } else {
-                break;
-            }
-        }
-    }
-    return (pow);
+    return std::pow(a, b);
 }
 //=============================================================================
 double
 power_dd(double a, double b)
 {
+    if (b == 0) {
+        return a;
+    }
+    if (b == 1) {
+        return a;
+    }
+    if (std::isnan(a)) {
+        return std::nan("NaN");
+    }
+    if (std::isnan(b)) {
+        return std::nan("NaN");
+    }
+    if (a == 0) {
+        if (std::isinf(b) && b < 0) {
+            return std::numeric_limits<double>::infinity();
+        }
+        if (std::isinf(b) && b > 0) {
+            return 0;
+        }
+        if (b < 0) {
+            return std::numeric_limits<double>::infinity();
+        }
+        return 0;
+    }
+    if (a == -1 && std::isinf(b)) {
+        return 1;
+    }
+    if (std::isinf(b) && b > 0) {
+        if (abs(a) > 1) {
+            return std::numeric_limits<double>::infinity();
+        }
+    }
+    if (std::isinf(b) && b < 0) {
+        if (abs(a) > 1) {
+            return 0;
+        }
+    }
+    if (std::isinf(b) && b > 0) {
+        if (abs(a) < 1) {
+            return 0;
+        }
+    }
+    if (std::isinf(b) && b < 0) {
+        if (abs(a) < 1) {
+            return std::numeric_limits<double>::infinity();
+        }
+    }
+    if (std::isinf(a) && a > 0) {
+        if (b > 0) {
+            return std::numeric_limits<double>::infinity();
+        }
+        if (b < 0) {
+            return 0;
+        }
+    }
+    if (std::isinf(a) && a < 0) {
+        return pow(-0, -b);
+    }
+    if (a < 0 && std::trunc(b) == b) {
+        return std::nan("NaN");
+    }
     return pow(a, b);
 }
 //=============================================================================
 void
-cicpower(int n, float* c, float* a, int stride1, int* b, int stride2)
+cicpower(int n, single* c, single* a, int stride1, int* b, int stride2)
 {
     int m, p;
     double z1[2];
@@ -165,15 +143,15 @@ cicpower(int n, float* c, float* a, int stride1, int* b, int stride2)
         z1[0] = a[2 * m];
         z1[1] = a[2 * m + 1];
         power_zi(z3, z1, b[p]);
-        c[2 * i] = (float)z3[0];
-        c[2 * i + 1] = (float)z3[1];
+        c[2 * i] = (single)z3[0];
+        c[2 * i + 1] = (single)z3[1];
         m += stride1;
         p += stride2;
     }
 }
 //=============================================================================
 void
-cfcpower(int n, float* c, float* a, int stride1, float* b, int stride2)
+cfcpower(int n, single* c, single* a, int stride1, single* b, int stride2)
 {
     int m, p;
     double z1[2];
@@ -187,8 +165,8 @@ cfcpower(int n, float* c, float* a, int stride1, float* b, int stride2)
         z2[0] = b[p];
         z2[1] = 0;
         power_zz(z3, z1, z2);
-        c[2 * i] = (float)z3[0];
-        c[2 * i + 1] = (float)z3[1];
+        c[2 * i] = (single)z3[0];
+        c[2 * i + 1] = (single)z3[1];
         m += stride1;
         p += stride2;
     }
@@ -217,7 +195,7 @@ zdzpower(int n, double* c, double* a, int stride1, double* b, int stride2)
 }
 //=============================================================================
 void
-cccpower(int n, float* c, float* a, int stride1, float* b, int stride2)
+cccpower(int n, single* c, single* a, int stride1, single* b, int stride2)
 {
     int m, p;
     double z1[2];
@@ -231,8 +209,8 @@ cccpower(int n, float* c, float* a, int stride1, float* b, int stride2)
         z2[0] = b[2 * p];
         z2[1] = b[2 * p + 1];
         power_zz(z3, z1, z2);
-        c[2 * i] = (float)z3[0];
-        c[2 * i + 1] = (float)z3[1];
+        c[2 * i] = (single)z3[0];
+        c[2 * i + 1] = (single)z3[1];
         m += stride1;
         p += stride2;
     }
@@ -282,9 +260,7 @@ zizpower(int n, double* c, double* a, int stride1, int* b, int stride2)
 void
 didpower(int n, double* c, double* a, int stride1, int* b, int stride2)
 {
-    int m, p;
-    m = 0;
-    p = 0;
+    int m = 0, p = 0;
     for (int i = 0; i < n; i++) {
         c[i] = power_di(a[m], b[p]);
         m += stride1;
@@ -295,9 +271,7 @@ didpower(int n, double* c, double* a, int stride1, int* b, int stride2)
 void
 dddpower(int n, double* c, double* a, int stride1, double* b, int stride2)
 {
-    int m, p;
-    m = 0;
-    p = 0;
+    int m = 0, p = 0;
     for (int i = 0; i < n; i++) {
         c[i] = power_dd(a[m], b[p]);
         m += stride1;
@@ -306,51 +280,29 @@ dddpower(int n, double* c, double* a, int stride1, double* b, int stride2)
 }
 //=============================================================================
 void
-fifpower(int n, float* c, float* a, int stride1, int* b, int stride2)
+fifpower(int n, single* c, single* a, int stride1, int* b, int stride2)
 {
-    int m, p;
-    m = 0;
-    p = 0;
+    int m = 0, p = 0;
     for (int i = 0; i < n; i++) {
-        c[i] = (float)power_di(a[m], b[p]);
+        c[i] = (single)power_di(a[m], b[p]);
         m += stride1;
         p += stride2;
     }
 }
 //=============================================================================
 void
-fffpower(int n, float* c, float* a, int stride1, float* b, int stride2)
+fffpower(int n, single* c, single* a, int stride1, single* b, int stride2)
 {
-    int m, p;
-    m = 0;
-    p = 0;
+    int m = 0, p = 0;
     for (int i = 0; i < n; i++) {
-        c[i] = (float)power_dd(a[m], b[p]);
+        c[i] = (single)power_dd(a[m], b[p]);
         m += stride1;
         p += stride2;
     }
 }
 //=============================================================================
-/**
- * This is the generic function interface into calculations
- * that can be performed on some type.
- */
 typedef void (*vvfun)(indexType length, void* result, const void* arg1, const int stride1,
     const void* arg2, const int stride2);
-//=============================================================================
-/**
- * This structure holds pointers to functions when
- * all types can be handled.
- */
-/*
-typedef struct {
-    vvfun int32func;
-    vvfun floatfunc;
-    vvfun doublefunc;
-    vvfun complexfunc;
-    vvfun dcomplexfunc;
-} packVectorVector;
-*/
 //=============================================================================
 inline ArrayOf
 doPowerAssist(ArrayOf A, Class AClass, ArrayOf B, Class BClass, Class CClass, vvfun exec)
@@ -361,19 +313,19 @@ doPowerAssist(ArrayOf A, Class AClass, ArrayOf B, Class BClass, Class CClass, vv
     if (A.isScalar()) {
         indexType Blen(B.getLength());
         C = ArrayOf(CClass, B.getDimensions(), NULL);
-        void* Cp = C.allocateArrayOf(CClass, Blen);
+        void* Cp = ArrayOf::allocateArrayOf(CClass, Blen * C.getElementSize());
         exec(Blen, Cp, A.getDataPointer(), 0, B.getDataPointer(), 1);
         C.setDataPointer(Cp);
     } else if (B.isScalar()) {
         indexType Alen(A.getLength());
         C = ArrayOf(CClass, A.getDimensions(), NULL);
-        void* Cp = C.allocateArrayOf(CClass, Alen);
+        void* Cp = ArrayOf::allocateArrayOf(CClass, Alen * C.getElementSize());
         exec(Alen, Cp, A.getDataPointer(), 1, B.getDataPointer(), 0);
         C.setDataPointer(Cp);
     } else {
         indexType Alen(A.getLength());
         C = ArrayOf(CClass, A.getDimensions(), NULL);
-        void* Cp = C.allocateArrayOf(CClass, Alen);
+        void* Cp = ArrayOf::allocateArrayOf(CClass, Alen * C.getElementSize());
         exec(Alen, Cp, A.getDataPointer(), 1, B.getDataPointer(), 1);
         C.setDataPointer(Cp);
     }
@@ -387,6 +339,7 @@ doPowerAssist(ArrayOf A, Class AClass, ArrayOf B, Class BClass, Class CClass, vv
 #define MAPOP(o, a, b, c, f)                                                                       \
     case o:                                                                                        \
         return doPowerAssist(A, a, B, b, c, f);
+//=============================================================================
 inline ArrayOf
 DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
 {
@@ -395,31 +348,37 @@ DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
     stringVector dummySV;
     Class AClass, BClass;
     int opType;
+
     if (A.isEmpty() || B.isEmpty()) {
-        return ArrayOf::emptyConstructor();
+        if (A.isEmpty()) {
+            C = ArrayOf::emptyConstructor(A.getDimensions());
+        } else {
+            C = ArrayOf::emptyConstructor(B.getDimensions());
+        }
+        return C;
     }
     CheckNumeric(A, B, "^");
     Dimensions dimsA = A.getDimensions();
     Dimensions dimsB = B.getDimensions();
     if (!(SameSizeCheck(dimsA, dimsB) || A.isScalar() || B.isScalar())) {
-        Error(L"Size mismatch on arguments to power (^) operator.");
+        Error(_W("Size mismatch on arguments to power (^) operator."));
     }
-    // If A is not at least a float type, promote it to double
+    // If A is not at least a single type, promote it to double
     AClass = A.getDataClass();
     BClass = B.getDataClass();
-    if (AClass < NLS_SINGLE) {
+    if (AClass < NLS_SINGLE)
         AClass = NLS_DOUBLE;
-    }
-    if (BClass < NLS_INT64) {
+    if (BClass < NLS_INT64)
         BClass = NLS_INT64;
-    }
     // Get a read on if A is positive
     Anegative = !(A.isPositive());
     // Check through the different type cases...
     opType = 0;
     if (AClass == NLS_SCOMPLEX) {
         switch (BClass) {
-            OPCASE(NLS_INT32, 1);
+        default:
+            Error(_W("Unhandled type for second argument to A^B"));
+            OPCASE(NLS_INT64, 1);
             OPCASE(NLS_SINGLE, 2);
             OPCASE(NLS_DOUBLE, 3);
             OPCASE(NLS_SCOMPLEX, 4);
@@ -427,7 +386,9 @@ DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
         }
     } else if (AClass == NLS_DCOMPLEX) {
         switch (BClass) {
-            OPCASE(NLS_INT32, 6);
+        default:
+            Error(_W("Unhandled type for second argument to A^B"));
+            OPCASE(NLS_INT64, 6);
             OPCASE(NLS_SINGLE, 3);
             OPCASE(NLS_DOUBLE, 3);
             OPCASE(NLS_SCOMPLEX, 5);
@@ -435,7 +396,9 @@ DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
         }
     } else if (AClass == NLS_DOUBLE && Anegative) {
         switch (BClass) {
-            OPCASE(NLS_INT32, 7);
+        default:
+            Error(_W("Unhandled type for second argument to A^B"));
+            OPCASE(NLS_INT64, 7);
             OPCASE(NLS_SINGLE, 5);
             OPCASE(NLS_DOUBLE, 5);
             OPCASE(NLS_SCOMPLEX, 5);
@@ -443,7 +406,9 @@ DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
         }
     } else if (AClass == NLS_DOUBLE && (!Anegative)) {
         switch (BClass) {
-            OPCASE(NLS_INT32, 7);
+        default:
+            Error(_("Unhandled type for second argument to A^B"));
+            OPCASE(NLS_INT64, 7);
             OPCASE(NLS_SINGLE, 8);
             OPCASE(NLS_DOUBLE, 8);
             OPCASE(NLS_SCOMPLEX, 5);
@@ -451,7 +416,9 @@ DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
         }
     } else if (AClass == NLS_SINGLE && Anegative) {
         switch (BClass) {
-            OPCASE(NLS_INT32, 9);
+        default:
+            Error(_("Unhandled type for second argument to A^B"));
+            OPCASE(NLS_INT64, 9);
             OPCASE(NLS_SINGLE, 4);
             OPCASE(NLS_DOUBLE, 5);
             OPCASE(NLS_SCOMPLEX, 4);
@@ -459,7 +426,9 @@ DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
         }
     } else if (AClass == NLS_SINGLE && (!Anegative)) {
         switch (BClass) {
-            OPCASE(NLS_INT32, 9);
+        default:
+            Error(_("Unhandled type for second argument to A^B"));
+            OPCASE(NLS_INT64, 9);
             OPCASE(NLS_SINGLE, 10);
             OPCASE(NLS_DOUBLE, 8);
             OPCASE(NLS_SCOMPLEX, 4);
@@ -468,24 +437,86 @@ DoPowerTwoArgFunction(ArrayOf A, ArrayOf B)
     }
     // Invoke the appropriate case
     switch (opType) {
-        MAPOP(1, NLS_SCOMPLEX, NLS_INT32, NLS_SCOMPLEX, (vvfun)cicpower);
+    default:
+        Error(_W("Unhandled type combination for A^B"));
+        MAPOP(1, NLS_SCOMPLEX, NLS_INT64, NLS_SCOMPLEX, (vvfun)cicpower);
         MAPOP(2, NLS_SCOMPLEX, NLS_SINGLE, NLS_SCOMPLEX, (vvfun)cfcpower);
         MAPOP(3, NLS_DCOMPLEX, NLS_DOUBLE, NLS_DCOMPLEX, (vvfun)zdzpower);
         MAPOP(4, NLS_SCOMPLEX, NLS_SCOMPLEX, NLS_SCOMPLEX, (vvfun)cccpower);
         MAPOP(5, NLS_DCOMPLEX, NLS_DCOMPLEX, NLS_DCOMPLEX, (vvfun)zzzpower);
-        MAPOP(6, NLS_DCOMPLEX, NLS_INT32, NLS_DCOMPLEX, (vvfun)zizpower);
-        MAPOP(7, NLS_DOUBLE, NLS_INT32, NLS_DOUBLE, (vvfun)didpower);
+        MAPOP(6, NLS_DCOMPLEX, NLS_INT64, NLS_DCOMPLEX, (vvfun)zizpower);
+        MAPOP(7, NLS_DOUBLE, NLS_INT64, NLS_DOUBLE, (vvfun)didpower);
         MAPOP(8, NLS_DOUBLE, NLS_DOUBLE, NLS_DOUBLE, (vvfun)dddpower);
-        MAPOP(9, NLS_SINGLE, NLS_INT32, NLS_SINGLE, (vvfun)fifpower);
+        MAPOP(9, NLS_SINGLE, NLS_INT64, NLS_SINGLE, (vvfun)fifpower);
         MAPOP(10, NLS_SINGLE, NLS_SINGLE, NLS_SINGLE, (vvfun)fffpower);
     }
     return ArrayOf();
 }
 //=============================================================================
 ArrayOf
-DotPower(ArrayOf A, ArrayOf B)
+DotPower(ArrayOf& A, ArrayOf& B, bool& needToOverload)
 {
-    return (DoPowerTwoArgFunction(A, B));
+    needToOverload = false;
+    if (A.isSparse() || B.isSparse()) {
+        needToOverload = true;
+        return ArrayOf();
+    }
+    Class destinationClass;
+    if (A.getDataClass() != B.getDataClass()) {
+        if ((A.isDoubleClass() || A.isSingleClass()) && (B.isDoubleClass() || B.isSingleClass())) {
+            if (A.isComplex() || B.isComplex()) {
+                destinationClass = NLS_DCOMPLEX;
+                A.promoteType(NLS_DCOMPLEX);
+                B.promoteType(NLS_DCOMPLEX);
+            } else {
+                destinationClass = NLS_DOUBLE;
+                A.promoteType(NLS_DOUBLE);
+                B.promoteType(NLS_DOUBLE);
+            }
+        } else {
+            needToOverload = true;
+            return ArrayOf();
+        }
+    } else {
+        switch (A.getDataClass()) {
+        case NLS_UINT8:
+        case NLS_INT8:
+        case NLS_UINT16:
+        case NLS_INT16:
+        case NLS_UINT32:
+        case NLS_INT32:
+        case NLS_UINT64:
+        case NLS_INT64:
+        case NLS_SINGLE:
+        case NLS_DOUBLE:
+        case NLS_SCOMPLEX:
+        case NLS_DCOMPLEX:
+        case NLS_CHAR: {
+            destinationClass = A.getDataClass();
+        } break;
+        case NLS_LOGICAL:
+        case NLS_HANDLE:
+        case NLS_CELL_ARRAY:
+        case NLS_STRUCT_ARRAY:
+        default: {
+            needToOverload = true;
+            return ArrayOf();
+        } break;
+        }
+    }
+    ArrayOf res = DoPowerTwoArgFunction(A, B);
+    if (res.getDataClass() == NLS_DCOMPLEX) {
+        if (res.allReal()) {
+            res.promoteType(NLS_DOUBLE);
+        }
+    } else if (res.getDataClass() == NLS_SCOMPLEX) {
+        if (res.allReal()) {
+            res.promoteType(NLS_DOUBLE);
+        }
+    } else {
+        res.promoteType(destinationClass);
+    }
+    return res;
 }
 //=============================================================================
 }