+++ modules/elementary_functions/src/include/DotPower.hpp
@@ -22,10 +22,12 @@
 #include "nlsElementary_functions_exports.h"
 //=============================================================================
 namespace Nelson {
+//=============================================================================
 /**
  * Element-wise power.  A .^ B
  */
 NLSELEMENTARY_FUNCTIONS_IMPEXP ArrayOf
-DotPower(ArrayOf A, ArrayOf B);
+DotPower(ArrayOf& A, ArrayOf& B, bool& needToOverload);
+//=============================================================================
 } // namespace Nelson
 //=============================================================================